+++ recipe.json
@@ -1,1 +1,1 @@
-[{"id":3,"name":"XXX","ingredients":["2","3","4"],"description":"default","is_favorite":false},{"id":4,"name":"XXX","ingredients":["1","2","3"],"description":"default","is_favorite":false},{"id":7,"name":"123","ingredients":["1","2","3","4","5","6","7"],"description":"asfopsdfio dffiojsdiojf sdgfsdfio","is_favorite":false},{"id":8,"name":"TEST","ingredients":["apple","milk","sho","chocolate"],"description":"Drink water","is_favorite":true},{"id":9,"name":"Chicken wit potato","ingredients":["salt","potato","chicken","oil"],"description":"boli chicken,add salt, bla bla bla","is_favorite":true},{"id":10,"name":"Северин","ingredients":["табір","церква"],"description":"Северин пішов у табір а попав у церкву","is_favorite":false},{"id":11,"name":"вфі","ingredients":["вфі"],"description":"фві","is_favorite":false}]
+[{"id":3,"name":"XXX","ingredients":["2","3","4"],"description":"default","is_favorite":false},{"id":4,"name":"XXX","ingredients":["1","2","3"],"description":"default","is_favorite":false},{"id":7,"name":"123","ingredients":["1","2","3","4","5","6","7"],"description":"asfopsdfio dffiojsdiojf sdgfsdfio","is_favorite":false},{"id":8,"name":"TEST","ingredients":["apple","milk","sho","chocolate"],"description":"Drink water","is_favorite":true},{"id":9,"name":"Chicken wit potato","ingredients":["salt","potato","chicken","oil"],"description":"boli chicken,add salt, bla bla bla","is_favorite":true},{"id":10,"name":"Северин","ingredients":["табір","церква"],"description":"Северин пішов у табір а попав у церкву","is_favorite":false},{"id":11,"name":"вфі","ingredients":["вфі"],"description":"фві","is_favorite":false},{"id":8,"name":"Bizeps","ingredients":["Viktor Ivanovich","Shtanga","Metandrostenalon"],"description":"Just do this 3 things together, and you will become like big like Ihor! ","is_favorite":false}]
